move ReadHandler in its own file
