+++ dnp3/src/master/mod.rs
@@ -3,12 +3,14 @@ pub use error::*;
 pub use file::*;
 pub use handler::*;
 pub use poll::PollHandle;
+pub use read_handler::*;
 pub use request::*;
 
 mod association;
 mod error;
 mod file;
 mod handler;
+mod read_handler;
 mod request;
 
 pub(crate) mod convert;
